Solved Container with Most Water
Implemented a two-pointer approach to solve the 'Container With Most Water' problem.

Notable lessons from this were again to be extra mindful with if/elif/else statements. It is very easy to introduce unnecessary evaluations if not very mindful of the logic of the problem. This may not break the program, but it will definitely create inefficiencies.