Solved Trapping Rain Water
Solved the Trapping Rain Water problem using O(n) time and space complexity.

2 solutions have been included - will be adding a 3rd. Both intro solutions work, but are not completely optimal.

Important lessons from this are again being very mindful of array management. It is incredibly easy to use existing methods to make up for flawed logic. Only use these tools when it is necessary to do so - avoiding append/insert/reverse is good if not mandatory to use.