3 Binary Search Problems
Completed 3 binary Search problems (Koko eating bananas, find_min() of a rotated sorted array, and search in a rotated sorted array.

Each uses a rather straightforward implementation of BS and each runs in optimal runtime.

One worthwhile note was in searching the rotated array. I had to use BS to find a pivot (the min) of the sorted array, after doing so i was able to optimize search